+++ nextgisweb/wfsserver/wfs_handler.py
@@ -3,15 +3,19 @@ from __future__ import division, absolute_import, print_function, unicode_litera
 
 from collections import OrderedDict
 from datetime import datetime
-from lxml import etree
-from lxml.builder import ElementMaker
+from os import path
+from tempfile import NamedTemporaryFile
 
+from lxml import etree, html
+from lxml.builder import ElementMaker
 from osgeo import ogr, osr
+from pyramid.request import Request
 from six import BytesIO, text_type
 
 from ..core.exception import ValidationError
 from ..feature_layer import Feature, FIELD_TYPE, GEOM_TYPE
 from ..geometry import box, geom_from_wkb
+from ..layer import IBboxLayer
 from ..resource import DataScope
 from ..spatial_ref_sys import SRS
 
@@ -26,44 +30,46 @@ VERSION_SUPPORTED = (v100, v200, v202)
 
 VERSION_DEFAULT = v202
 
+XSD_DIR = path.join(path.dirname(
+    path.abspath(__file__)), 'test/xsd/')
 
 _nsmap = dict(
     wfs=OrderedDict((
-        (v100, 'http://www.opengis.net/wfs'),
-        (v200, 'http://www.opengis.net/wfs/2.0'),
+        (v100, ('http://www.opengis.net/wfs', 'http://schemas.opengis.net/wfs/1.0.0/WFS-basic.xsd')),
+        (v200, ('http://www.opengis.net/wfs/2.0', 'http://schemas.opengis.net/wfs/2.0/wfs.xsd')),
     )),
     gml=OrderedDict((
-        (v100, 'http://www.opengis.net/gml'),
-        (v200, 'http://www.opengis.net/gml/3.2'),
+        (v100, ('http://www.opengis.net/gml', 'http://schemas.opengis.net/gml/2.1.2/feature.xsd')),
+        (v200, ('http://www.opengis.net/gml/3.2', 'http://schemas.opengis.net/gml/3.2.1/feature.xsd')),
     )),
     ogc=OrderedDict((
-        (v100, 'http://www.opengis.net/ogc'),
+        (v100, ('http://www.opengis.net/ogc', None)),
     )),
     xsi=OrderedDict((
-        (v100, 'http://www.w3.org/2001/XMLSchema-instance'),
+        (v100, ('http://www.w3.org/2001/XMLSchema-instance', None)),
     )),
     ows=OrderedDict((
-        (v200, 'http://www.opengis.net/ows/1.1'),
+        (v200, ('http://www.opengis.net/ows/1.1', None)),
     )),
     xlink=OrderedDict((
-        (v200, 'http://www.w3.org/1999/xlink'),
+        (v200, ('http://www.w3.org/1999/xlink', None)),
     )),
     fes=OrderedDict((
-        (v200, 'http://www.opengis.net/fes/2.0'),
+        (v200, ('http://www.opengis.net/fes/2.0', None)),
     ))
 )
 
 
 def nsmap(prefix, request_version):
     item = _nsmap[prefix]
-    for version, value in reversed(item.items()):
+    for version, (ns, loc) in reversed(item.items()):
         if version <= request_version:
-            return value
-    return None
+            return dict(ns=ns, loc=loc)
+    raise ValidationError("Namespace %s not found for version %s." % (prefix, request_version))
 
 
 def ns_attr(ns, attr, request_version):
-    return '{{{0}}}{1}'.format(nsmap(ns, request_version), attr)
+    return '{{{0}}}{1}'.format(nsmap(ns, request_version)['ns'], attr)
 
 
 def ns_trim(value):
@@ -82,6 +88,18 @@ def El(tag, attrs=None, parent=None, text=None, namespace=None):
     if parent is not None:
         parent.append(e)
     return e
+
+
+def find_tags(element, tag):
+    return element.xpath('.//*[local-name()="%s"]' % tag)
+
+
+def fid_encode(fid):
+    return 'id-' + str(fid)
+
+
+def fid_decode(fid):
+    return int(fid[3:])
 
 
 GET_CAPABILITIES = 'GetCapabilities'
@@ -116,14 +134,15 @@ def geom_from_gml(el):
 
 
 class WFSHandler():
-    def __init__(self, resource, request):
+    def __init__(self, resource, request, force_schema_validation=False):
         self.resource = resource
         self.request = request
 
         if self.request.method == 'GET':
             params = request.params
         elif self.request.method == 'POST':
-            self.root_body = etree.parse(BytesIO(self.request.body)).getroot()
+            parser = etree.XMLParser(recover=True)
+            self.root_body = etree.parse(BytesIO(self.request.body), parser=parser).getroot()
             params = self.root_body.attrib
         else:
             raise ValidationError("Unsupported request method")
@@ -131,7 +150,7 @@ class WFSHandler():
         # 6.2.5.2 Parameter names shall not be case sensitive
         params = dict((k.upper(), v) for k, v in params.items())
 
-        self.p_requset = params.get('REQUEST') if self.request.method == 'GET' \
+        self.p_request = params.get('REQUEST') if self.request.method == 'GET' \
             else ns_trim(self.root_body.tag)
 
         av = params.get('ACCEPTVERSIONS')
@@ -157,37 +176,97 @@ class WFSHandler():
         self.p_count = params.get('COUNT', params.get('MAXFEATURES'))
         self.p_startindex = params.get('STARTINDEX')
 
+        self.p_validate_schema = force_schema_validation or (
+            params.get('VALIDATESCHEMA', 'FALSE').upper() in ('1', 'YES', 'TRUE'))
+
+        self.service_namespace = self.request.route_url(
+            'wfsserver.wfs', id=self.resource.id)
+
+    @property
+    def title(self):
+        return self.resource.display_name
+
+    @property
+    def abstract(self):
+        return html.document_fromstring(self.resource.description).text_content() \
+            if self.resource.description is not None else ''
+
     @property
     def gml_format(self):
-        return 'GML3' if self.p_version >= v110 else 'GML2'
+        return 'GML32' if self.p_version >= v110 else 'GML2'
 
     def response(self):
-        if self.p_requset == GET_CAPABILITIES:
-            return self._get_capabilities200() if self.p_version >= v200 \
+        if self.p_request == GET_CAPABILITIES:
+            xml = self._get_capabilities200() if self.p_version >= v200 \
                 else self._get_capabilities()
-        elif self.p_requset == DESCRIBE_FEATURE_TYPE:
-            return self._describe_feature_type()
-        elif self.p_requset == GET_FEATURE:
-            return self._get_feature()
-        elif self.p_requset == TRANSACTION:
-            return self._transaction()
+        elif self.p_request == DESCRIBE_FEATURE_TYPE:
+            xml = self._describe_feature_type()
+        elif self.p_request == GET_FEATURE:
+            xml = self._get_feature()
+        elif self.p_request == TRANSACTION:
+            xml = self._transaction()
         else:
             raise ValidationError("Unsupported request")
 
-    def _feature_type_list(self, parent):
-        _ns_ows = nsmap('ows', self.p_version)
+        if self.p_validate_schema:
+            if self.p_request in (GET_CAPABILITIES, TRANSACTION):
+                version_dir = '1.0.0' if self.p_version == v100 else '2.0'
+                wfs_schema_dir = path.join(XSD_DIR, 'schemas.opengis.net/wfs/')
+                xsd_file = 'WFS-capabilities.xsd' if self.p_version == v100 \
+                    and self.p_request == GET_CAPABILITIES else 'wfs.xsd'
+                xsd_path = path.join(wfs_schema_dir, version_dir, xsd_file)
 
+                schema = etree.XMLSchema(file=xsd_path)
+                schema.assertValid(etree.XML(xml))
+            elif self.p_request == DESCRIBE_FEATURE_TYPE:
+                xsd_path = path.join(XSD_DIR, 'www.w3.org/2009/XMLSchema/XMLSchema.xsd')
+
+                schema = etree.XMLSchema(file=xsd_path)
+                schema.assertValid(etree.XML(xml))
+            elif self.p_request == GET_FEATURE:
+                describe_path = self.request.route_path(
+                    'wfsserver.wfs', id=self.resource.id, _query=dict(
+                        REQUEST=DESCRIBE_FEATURE_TYPE, SERVICE='WFS',
+                        VERSION=self.p_version, TYPENAME=self.p_typenames))
+                subreq = Request.blank(describe_path)
+                subreq.headers = self.request.headers
+                resp = self.request.invoke_subrequest(subreq)
+                describe_root = etree.XML(resp.body)
+
+                opengis_url = 'http://schemas.opengis.net'
+                opengis_dir = path.join(XSD_DIR, 'schemas.opengis.net')
+                for el in describe_root.xpath('.//*[starts-with(@schemaLocation, \'%s\')]' % opengis_url):
+                    el.attrib['schemaLocation'] = el.attrib['schemaLocation'].replace(opengis_url, opengis_dir)
+
+                with NamedTemporaryFile() as tmp:
+                    tmp.write(etree.tostring(describe_root))
+                    tmp.flush()
+
+                    _schema = El('schema', dict(elementFormDefault='qualified'), namespace='http://www.w3.org/2001/XMLSchema')
+                    El('import', dict(namespace=self.service_namespace, schemaLocation=tmp.name), parent=_schema, namespace='http://www.w3.org/2001/XMLSchema')
+
+                    schema = etree.XMLSchema(etree=_schema)
+                    schema.assertValid(etree.XML(xml))
+            else:
+                raise ValidationError("Schema validation isn't supported for {} request".format(self.p_request))
+
+        return xml
+
+    def _feature_type_list(self, parent):
         __list = El('FeatureTypeList', parent=parent)
         if self.p_version < v200:
             __ops = El('Operations', parent=__list)
             El('Query', parent=__ops)
 
+        EM_name = ElementMaker(nsmap=dict(ngw=self.service_namespace))
         for layer in self.resource.layers:
             feature_layer = layer.resource
             if not feature_layer.has_permission(DataScope.read, self.request.user):
                 continue
             __type = El('FeatureType', parent=__list)
-            El('Name', parent=__type, text=layer.keyname)
+            __name = EM_name('Name')
+            __name.text = layer.keyname
+            __type.append(__name)
             El('Title', parent=__type, text=layer.display_name)
             El('Abstract', parent=__type)
             if self.p_version >= v200:
@@ -205,29 +284,47 @@ class WFSHandler():
                     El('Update', parent=__ops)
                     El('Delete', parent=__ops)
 
-            extent = feature_layer.extent
-            if self.p_version >= v110:
-                __bbox = El('WGS84BoundingBox', namespace=_ns_ows, parent=__type)
-                El('LowerCorner', namespace=_ns_ows, parent=__bbox,
-                   text='%.6f %.6f' % (extent['minLon'], extent['minLat']))
-                El('UpperCorner', namespace=_ns_ows, parent=__bbox,
-                   text='%.6f %.6f' % (extent['maxLon'], extent['maxLat']))
+            if IBboxLayer.providedBy(feature_layer):
+                extent = feature_layer.extent
+                if self.p_version >= v110:
+                    _ns_ows = nsmap('ows', self.p_version)['ns']
+                    __bbox = El('WGS84BoundingBox', namespace=_ns_ows, parent=__type)
+                    El('LowerCorner', namespace=_ns_ows, parent=__bbox,
+                       text='%.6f %.6f' % (extent['minLon'], extent['minLat']))
+                    El('UpperCorner', namespace=_ns_ows, parent=__bbox,
+                       text='%.6f %.6f' % (extent['maxLon'], extent['maxLat']))
+                else:
+                    bbox = dict(maxx=str(extent['maxLon']), maxy=str(extent['maxLat']),
+                                minx=str(extent['minLon']), miny=str(extent['minLat']))
+                    El('LatLongBoundingBox', bbox, parent=__type)
+
+    def _parse_filter(self, __filter):
+        v_gt200 = self.p_version >= v200
+        resid_tag = 'ResourceId' if v_gt200 else 'FeatureId'
+        resid_attr = 'rid' if v_gt200 else 'fid'
+
+        fid = None
+        for __el in __filter:
+            if ns_trim(__el.tag) == resid_tag:
+                if fid is not None:
+                    raise ValueError("Multiple feature ID filter not supported.")
+                else:
+                    fid = fid_decode(__el.get(resid_attr))
             else:
-                bbox = dict(maxx=str(extent['maxLon']), maxy=str(extent['maxLat']),
-                            minx=str(extent['minLon']), miny=str(extent['minLat']))
-                El('LatLongBoundingBox', bbox, parent=__type)
+                raise ValueError("Filter element '%s' not supported." % __el.tag)
+        return fid
 
     def _get_capabilities(self):
-        EM = ElementMaker(nsmap=dict(ogc=nsmap('ogc', self.p_version)))
+        EM = ElementMaker(nsmap=dict(ogc=nsmap('ogc', self.p_version)['ns']))
         root = EM('WFS_Capabilities', dict(
             version=self.p_version,
-            xmlns=nsmap('wfs', self.p_version)))
+            xmlns=nsmap('wfs', self.p_version)['ns']))
 
         # Service
         __s = El('Service', parent=root)
-        El('Name', parent=__s, text='WFS Server')
-        El('Title', parent=__s, text='Web Feature Service Server')
-        El('Abstract', parent=__s, text='Supports WFS')
+        El('Name', parent=__s, text=self.resource.keyname or 'WFS')
+        El('Title', parent=__s, text=self.title)
+        El('Abstract', parent=__s, text=self.abstract)
         El('OnlineResource', parent=__s)
 
         # Operations
@@ -257,7 +354,7 @@ class WFSHandler():
         self._feature_type_list(root)
 
         # Filter_Capabilities
-        _ns_ogc = nsmap('ogc', self.p_version)
+        _ns_ogc = nsmap('ogc', self.p_version)['ns']
         __filter = El('Filter_Capabilities', namespace=_ns_ogc, parent=root)
 
         __sc = El('Spatial_Capabilities', namespace=_ns_ogc, parent=__filter)
@@ -270,20 +367,20 @@ class WFSHandler():
         return etree.tostring(root)
 
     def _get_capabilities200(self):
-        _ns_ows = nsmap('ows', self.p_version)
-        _ns_fes = nsmap('fes', self.p_version)
+        _ns_ows = nsmap('ows', self.p_version)['ns']
+        _ns_fes = nsmap('fes', self.p_version)['ns']
 
         EM = ElementMaker(nsmap=dict(
-            fes=_ns_fes, ows=_ns_ows, xlink=nsmap('xlink', self.p_version)
+            fes=_ns_fes, ows=_ns_ows, xlink=nsmap('xlink', self.p_version)['ns']
         ))
         root = EM('WFS_Capabilities', dict(
             version=self.p_version,
-            xmlns=nsmap('wfs', self.p_version)))
+            xmlns=nsmap('wfs', self.p_version)['ns']))
 
         # Service
         __service = El('ServiceIdentification', namespace=_ns_ows, parent=root)
-        El('Title', namespace=_ns_ows, parent=__service, text='Web Feature Service Server')
-        El('Abstract', namespace=_ns_ows, parent=__service, text='Supports WFS')
+        El('Title', namespace=_ns_ows, parent=__service, text=self.title)
+        El('Abstract', namespace=_ns_ows, parent=__service, text=self.abstract)
         El('ServiceType', namespace=_ns_ows, parent=__service, text='WFS')
         for version in VERSION_SUPPORTED:
             El('ServiceTypeVersion', namespace=_ns_ows, parent=__service, text=version)
@@ -341,75 +438,113 @@ class WFSHandler():
         return etree.tostring(root)
 
     def _describe_feature_type(self):
-        _ns_gml = nsmap('gml', self.p_version)
+        gml = nsmap('gml', self.p_version)
+        wfs = nsmap('wfs', self.p_version)
 
-        EM = ElementMaker(nsmap=dict(gml=_ns_gml))
+        EM = ElementMaker(nsmap=dict(gml=gml['ns'], ngw=self.service_namespace))
         root = EM('schema', dict(
-            targetNamespace=nsmap('wfs', self.p_version),
+            targetNamespace=self.service_namespace,
             elementFormDefault='qualified',
             attributeFormDefault='unqualified',
             version='0.1',
             xmlns='http://www.w3.org/2001/XMLSchema'))
 
         El('import', dict(
-            namespace=_ns_gml,
-            schemaLocation='http://schemas.opengis.net/gml/2.0.0/feature.xsd'
+            namespace=gml['ns'],
+            schemaLocation=gml['loc']
+        ), parent=root)
+        El('import', dict(
+            namespace=wfs['ns'],
+            schemaLocation=wfs['loc']
         ), parent=root)
 
-        typenames = [layer.keyname for layer in self.resource.layers] \
-            if self.p_typenames is None else self.p_typenames.split(',')
+        if self.request.method == 'GET':
+            typenames = None if self.p_typenames is None \
+                else self.p_typenames.split(',')
+        elif self.request.method == 'POST':
+            __typenames = find_tags(self.root_body, 'TypeName')
+            typenames = None if len(__typenames) == 0 \
+                else [__typename.text for __typename in __typenames]
 
-        if len(typenames) == 1:
-            layer = Layer.filter_by(service_id=self.resource.id, keyname=typenames[0]).one()
-            El('element', dict(name=layer.keyname, substitutionGroup='gml:_Feature',
-                               type='%s_Type' % layer.keyname), parent=root)
-            __ctype = El('complexType', dict(name="%s_Type" % layer.keyname), parent=root)
+        if typenames is None:
+            typenames = [layer.keyname for layer in self.resource.layers]
+
+        for typename in typenames:
+            substitutionGroup = 'gml:AbstractFeature' if self.p_version > v100 else 'gml:_Feature'
+            El('element', dict(name=typename, substitutionGroup=substitutionGroup,
+                               type='ngw:%s_Type' % typename), parent=root)
+
+        for typename in typenames:
+            layer = Layer.filter_by(service_id=self.resource.id, keyname=typename).one()
+            feature_layer = layer.resource
+            __ctype = El('complexType', dict(name="%s_Type" % typename), parent=root)
             __ccontent = El('complexContent', parent=__ctype)
             __ext = El('extension', dict(base='gml:AbstractFeatureType'), parent=__ccontent)
             __seq = El('sequence', parent=__ext)
-            for field in layer.resource.fields:
+
+            if feature_layer.geometry_type not in GEOM_TYPE_TO_GML_TYPE:
+                raise ValidationError("Geometry type not supported: %s"
+                                      % feature_layer.geometry_type)
+            El('element', dict(minOccurs='0', name='geom', type=GEOM_TYPE_TO_GML_TYPE[
+                feature_layer.geometry_type]), parent=__seq)
+
+            for field in feature_layer.fields:
                 if field.datatype == FIELD_TYPE.REAL:
                     datatype = 'double'
+                elif field.datatype == FIELD_TYPE.DATETIME:
+                    datatype = 'dateTime'
                 else:
                     datatype = field.datatype.lower()
-                El('element', dict(minOccurs='0', name=field.keyname, type=datatype), parent=__seq)
-
-            if layer.resource.geometry_type not in GEOM_TYPE_TO_GML_TYPE:
-                raise ValidationError("Geometry type not supported: %s"
-                                      % layer.resource.geometry_type)
-            El('element', dict(minOccurs='0', name='geom', type=GEOM_TYPE_TO_GML_TYPE[
-                layer.resource.geometry_type]), parent=__seq)
-        else:
-            for keyname in typenames:
-                import_url = self.request.route_url(
-                    'wfsserver.wfs', id=self.resource.id,
-                    _query=dict(REQUEST='DescribeFeatureType', TYPENAME=keyname))
-                El('import', dict(schemaLocation=import_url), parent=root)
+                El('element', dict(minOccurs='0', name=field.keyname, type=datatype, nillable='true'), parent=__seq)
 
         return etree.tostring(root)
 
     def _get_feature(self):
-        _ns_wfs = nsmap('wfs', self.p_version)
-        _ns_gml = nsmap('gml', self.p_version)
+        wfs = nsmap('wfs', self.p_version)
+        gml = nsmap('gml', self.p_version)
+
+        __query = None
+
+        if self.request.method == 'POST':
+            __queries = find_tags(self.root_body, 'Query')
+            if len(__queries) > 1:
+                raise ValidationError("Multiple queries not supported.")
+            __query = __queries[0]
+            for k, v in __query.attrib.items():
+                if k.upper() in ('TYPENAME', 'TYPENAMES'):
+                    self.p_typenames = v
+                    break
 
         layer = Layer.filter_by(service_id=self.resource.id, keyname=self.p_typenames).one()
         feature_layer = layer.resource
         self.request.resource_permission(DataScope.read, feature_layer)
 
-        EM = ElementMaker(namespace=_ns_wfs, nsmap=dict(
-            gml=_ns_gml, wfs=_ns_wfs,
-            ogc=nsmap('ogc', self.p_version), xsi=nsmap('xsi', self.p_version)
+        EM = ElementMaker(namespace=wfs['ns'], nsmap=dict(
+            gml=gml['ns'], wfs=wfs['ns'], ngw=self.service_namespace,
+            ogc=nsmap('ogc', self.p_version)['ns'], xsi=nsmap('xsi', self.p_version)['ns']
         ))
+        describe_location = self.request.route_url(
+            'wfsserver.wfs', id=self.resource.id,
+            _query=dict(REQUEST=DESCRIBE_FEATURE_TYPE, SERVICE='WFS',
+                        VERSION=self.p_version, TYPENAME=self.p_typenames))
         schema_location = ' '.join((
-            _ns_wfs,
-            _ns_gml,
-            'http://schemas.opengis.net/gml/3.2.1/gml.xsd',
-            'http://schemas.opengis.net/wfs/2.0.0/wfs.xsd' if self.p_version >= v200
-            else 'http://schemas.opengeospatial.net/wfs/1.0.0/WFS-basic.xsd'
+            wfs['ns'], wfs['loc'],
+            gml['ns'], gml['loc'],
+            self.service_namespace, describe_location
         ))
-        root = EM('FeatureCollection', {ns_attr('xsi', 'schemaLocation', self.p_version): schema_location})  # NOQA: E501
+        root = EM('FeatureCollection', {
+            'xmlns': self.service_namespace,
+            ns_attr('xsi', 'schemaLocation', self.p_version): schema_location
+        })
 
         query = feature_layer.feature_query()
+
+        if __query is not None:
+            __filters = find_tags(__query, 'Filter')
+            if len(__filters) == 1:
+                fid = self._parse_filter(__filters[0])
+                if fid is not None:
+                    query.filter_by(id=fid)
 
         def parse_srs(value):
             # 'urn:ogc:def:crs:EPSG::3857' -> 3857
@@ -446,29 +581,55 @@ class WFSHandler():
             osr_out = osr.SpatialReference()
             osr_out.ImportFromWkt(srs_out.wkt)
 
+            __boundedBy = El('boundedBy', parent=root, namespace=gml['ns'] if self.p_version == v100 else wfs['ns'])
+            minX = maxX = minY = maxY = None
+
             for feature in query():
-                feature_id = str(feature.id)
-                __member = El('featureMember', {ns_attr('gml', 'id', self.p_version): feature_id},
-                              parent=root, namespace=_ns_gml)
-                __feature = El(layer.keyname, dict(fid=feature_id), parent=__member)
+                feature_id = fid_encode(feature.id)
+                __member = El('featureMember', parent=root, namespace=gml['ns']) if self.p_version == v100 \
+                    else El('member', parent=root, namespace=wfs['ns'])
+                id_attr = 'fid' if self.p_version == v100 else ns_attr('gml', 'id', self.p_version)
+                __feature = El(layer.keyname, {id_attr: feature_id}, parent=__member)
 
                 geom = ogr.CreateGeometryFromWkb(feature.geom.wkb, osr_out)
-                gml = geom.ExportToGML(['FORMAT=%s' % self.gml_format, 'NAMESPACE_DECL=YES'])
+
+                _minX, _maxX, _minY, _maxY = geom.GetEnvelope()
+                minX = _minX if minX is None else min(minX, _minX)
+                minY = _minY if minY is None else min(minY, _minY)
+                maxX = _maxX if maxX is None else max(maxX, _maxX)
+                maxY = _maxY if maxY is None else max(maxY, _maxY)
+
+                geom_gml = geom.ExportToGML([
+                    'FORMAT=%s' % self.gml_format,
+                    'NAMESPACE_DECL=YES',
+                    'GMLID=geom-%s' % feature_id])
                 __geom = El('geom', parent=__feature)
-                __gml = etree.fromstring(gml)
+                __gml = etree.fromstring(geom_gml)
                 __geom.append(__gml)
 
-                for field in feature.fields:
-                    _field = El(field, parent=__feature)
-                    value = feature.fields[field]
+                for field in feature_layer.fields:
+                    _field = El(field.keyname, parent=__feature)
+                    value = feature.fields[field.keyname]
                     if value is not None:
-                        if not isinstance(value, text_type):
+                        if isinstance(value, datetime):
+                            value = value.isoformat()
+                        elif not isinstance(value, text_type):
                             value = str(value)
                         _field.text = value
                     else:
                         _field.set(ns_attr('xsi', 'nil', self.p_version), 'true')
 
                 count += 1
+
+            if None in (minX, minY, maxX, maxY):
+                El('null', parent=__boundedBy, namespace=gml['ns'], text='unknown')
+            elif self.p_version == v100:
+                _box = El('Box', dict(srsName='EPSG:%d' % srs_out.id), parent=__boundedBy, namespace=gml['ns'])
+                El('coordinates', parent=_box, namespace=gml['ns'], text='%f %f %f %f' % (minX, minY, maxX, maxY))
+            else:
+                _envelope = El('Envelope', dict(srsName='urn:ogc:def:crs:EPSG::%d' % srs_out.id), parent=__boundedBy, namespace=gml['ns'])
+                El('lowerCorner', parent=_envelope, namespace=gml['ns'], text='%f %f' % (minX, minY))
+                El('upperCorner', parent=_envelope, namespace=gml['ns'], text='%f %f' % (maxX, maxY))
 
             matched = count
 
@@ -484,10 +645,8 @@ class WFSHandler():
         return etree.tostring(root)
 
     def _transaction(self):
-        v_gt200 = self.p_version >= v200
-
-        _ns_wfs = nsmap('wfs', self.p_version)
-        _ns_ogc = nsmap('ogc', self.p_version)
+        _ns_wfs = nsmap('wfs', self.p_version)['ns']
+        _ns_ogc = nsmap('ogc', self.p_version)['ns']
 
         layers = dict()
 
@@ -500,14 +659,19 @@ class WFSHandler():
             return layers[keyname]
 
         EM = ElementMaker(namespace=_ns_wfs, nsmap=dict(
-            wfs=_ns_wfs, ogc=_ns_ogc, xsi=nsmap('xsi', self.p_version)))
-        _response = EM('TransactionResponse', dict(version='1.0.0'))
+            wfs=_ns_wfs, ogc=_ns_ogc, xsi=nsmap('xsi', self.p_version)['ns']))
+        _response = EM('WFS_TransactionResponse' if self.p_version == v100
+                       else 'TransactionResponse', dict(version=self.p_version))
 
-        _summary = El('TransactionSummary', namespace=_ns_wfs, parent=_response)
-        summary = dict(totalInserted=0, totalUpdated=0, totalDeleted=0)
+        show_summary = self.p_version >= v200
+
+        if show_summary:
+            _summary = El('TransactionSummary', namespace=_ns_wfs, parent=_response)
+            summary = dict(totalInserted=0, totalUpdated=0, totalDeleted=0)
 
         for _operation in self.root_body:
-            if _operation.tag == ns_attr('wfs', 'Insert', self.p_version):
+            operation_tag = ns_trim(_operation.tag)
+            if operation_tag == 'Insert':
                 _layer = _operation[0]
                 keyname = ns_trim(_layer.tag)
                 feature_layer = find_layer(keyname)
@@ -525,27 +689,29 @@ class WFSHandler():
 
                 fid = feature_layer.feature_create(feature)
 
-                _insert = El('InsertResult', namespace=_ns_wfs, parent=_response)
-                El('FeatureId', dict(fid=str(fid)), namespace=_ns_ogc, parent=_insert)
+                _insert = El('InsertResult' if self.p_version == v100 else 'InsertResults',
+                             namespace=_ns_wfs, parent=_response)
+                El('FeatureId', dict(fid=fid_encode(fid)), namespace=_ns_ogc, parent=_insert)
 
-                summary['totalInserted'] += 1
+                if show_summary:
+                    summary['totalInserted'] += 1
             else:
                 keyname = ns_trim(_operation.get('typeName'))
                 feature_layer = find_layer(keyname)
 
-                _filter = _operation.find(ns_attr('ogc', 'Filter', self.p_version))
-                resid_tag = 'ResourceId' if v_gt200 else 'FeatureId'
-                resid_attr = 'rid' if v_gt200 else 'fid'
-                _feature_id = _filter.find(ns_attr('ogc', resid_tag, self.p_version))
-                fid = int(_feature_id.get(resid_attr))
+                _filter = find_tags(_operation, 'Filter')[0]
+                fid = self._parse_filter(_filter)
+                if fid is None:
+                    raise ValueError("Feature ID filter must be specified.")
 
-                if _operation.tag == ns_attr('wfs', 'Update', self.p_version):
+                if operation_tag == 'Update':
                     query = feature_layer.feature_query()
                     query.filter_by(id=fid)
                     feature = query().one()
-                    for _property in _operation.findall(ns_attr('wfs', 'Property', self.p_version)):
-                        key = _property.find(ns_attr('wfs', 'Name', self.p_version)).text
-                        _value = _property.find(ns_attr('wfs', 'Value', self.p_version))
+                    for _property in find_tags(_operation, 'Property'):
+                        key = find_tags(_property, 'Name')[0].text
+                        _values = find_tags(_property, 'Value')
+                        _value = None if len(_values) == 0 else _values[0]
 
                         geom_column = get_geom_column(feature_layer)
 
@@ -562,19 +728,23 @@ class WFSHandler():
 
                     feature_layer.feature_put(feature)
 
-                    summary['totalUpdated'] += 1
-                elif _operation.tag == ns_attr('wfs', 'Delete', self.p_version):
+                    if show_summary:
+                        summary['totalUpdated'] += 1
+                elif operation_tag == 'Delete':
                     feature_layer.feature_delete(fid)
-                    summary['totalDeleted'] += 1
+                    if show_summary:
+                        summary['totalDeleted'] += 1
                 else:
-                    raise NotImplementedError()
+                    raise ValidationError("Unknown operation: %s" % operation_tag)
 
-        for param, value in summary.items():
-            if value > 0:
-                El(param, namespace=_ns_wfs, text=str(value), parent=_summary)
+        if show_summary:
+            for param, value in summary.items():
+                if value > 0:
+                    El(param, namespace=_ns_wfs, text=str(value), parent=_summary)
 
-        _result = El('TransactionResult', namespace=_ns_wfs, parent=_response)
-        _status = El('Status', namespace=_ns_wfs, parent=_result)
-        El('SUCCESS', namespace=_ns_wfs, parent=_status)
+        if self.p_version == v100:
+            _result = El('TransactionResult', namespace=_ns_wfs, parent=_response)
+            _status = El('Status', namespace=_ns_wfs, parent=_result)
+            El('SUCCESS', namespace=_ns_wfs, parent=_status)
 
         return etree.tostring(_response)